Criar ponto e método para adicionar novo

+++ src/main/java/br/com/avenuecode/grupo5/timeKeeper/daos/UsuarioDao.java
@@ -27,7 +27,7 @@ public class UsuarioDao
       manager.merge(usuario);
    }
 
-   public Usuario get(Integer id){
+   public Usuario get(long id){
       return manager.find(Usuario.class, id);
    }
